tip commit

+++ server/controllers/employee.js
@@ -7,20 +7,21 @@ import TransferModal from "../models/transfer";
 import generatorPassword from "generate-password";
 
 export const newEmployee=async (req,res,next)=>{
-    const password=generatorPassword.generate({
-        length: 8,
-        numbers: true
-    });
-    const employeeInfo = {
-        name: req.body.name,
-        surname: req.body.surname,
-        phone: req.body.phone,
-        email:req.body.email,
-        category:req.body.categoryId,
-        role:"employee",
-        password,
-    }
     try {
+        const password=generatorPassword.generate({
+            length: 8,
+            numbers: true
+        });
+        const employeeInfo = {
+            name: req.body.name,
+            surname: req.body.surname,
+            phone: req.body.phone,
+            email:req.body.email,
+            category:req.body.categoryId,
+            role:"employee",
+            password,
+        }
+
         if(await AdminModel.countDocuments({email:req.body.email}) >0 ) return next(new CustomError("Bu E-posta  zaten bir departmana tanımlı",400))
         const newEmployee = await AdminModel.create(employeeInfo)
 
@@ -36,14 +37,20 @@ export const newEmployee=async (req,res,next)=>{
 }
 
 export const getLoginEmployee=async (req,res,next)=>{
-    const employeeId=req.employee._id
-    const employee=await AdminModel.findById(employeeId)
-    employee.password=undefined
-    res.status(200).json({
-        success:true,
-        message:"Giriş yapan çalışanın bilgileri getirildi.",
-        employee
-    })
+
+    try{
+        const employeeId=req.employee._id
+        const employee=await AdminModel.findById(employeeId)
+        employee.password=undefined
+        res.status(200).json({
+            success:true,
+            message:"Giriş yapan çalışanın bilgileri getirildi.",
+            employee
+        })
+    }catch (err){
+        next(err)
+    }
+
 }
 
 
@@ -69,11 +76,15 @@ export const getEmployeesByCategoryId=async (req,res,next)=>{
 }
 export const deleteEmployee=async (req,res,next)=>{
     try{
+
+        if(req.params.employeeId===req.params.transferEmployeeId){
+            return next(new CustomError("Şİkayetleri silinen kişiye atayamazsınız.",400));
+        }
+
         const deleteEmployee=await AdminModel.findById(req.params.employeeId)
         const transferReportsEmployee=await AdminModel.findById(req.params.transferEmployeeId)
         await ReportModal.updateMany({employee:req.params.employeeId},{employee:req.params.transferEmployeeId})
         await TransferModal.updateMany({employee:req.params.employeeId},{employee:req.params.transferEmployeeId})
-
         const deletedEmployee= await AdminModel.findByIdAndDelete(req.params.employeeId)
         res.status(200).json({
             success:true,
